requirements.txt
certifi==2019.6.16
Django>=2.2.1
django_compressor==2.3
pytz==2019.2
sqlparse==0.3.0
tensorflow==1.14.0
keras==2.2.4
pandas==0.25.0
matplotlib==3.1.0
plotly==4.0.0
arrow==0.13.1
bs4==0.0.1
numexpr==2.6.9
py-cpuinfo==5.0.0
psutil==5.6.3
tushare==1.2.40
sklearn==0.0
gitpython==3.0.1
jieba==0.39
gensim==3.8.0
pyecharts==0.5.11
